+++ magik-checks/src/main/java/org/stevenlooman/sw/magik/CheckList.java
@@ -8,6 +8,7 @@ import org.stevenlooman.sw.magik.checks.FormattingCheck;
 import org.stevenlooman.sw.magik.checks.LhsRhsComparatorEqualCheck;
 import org.stevenlooman.sw.magik.checks.LineLengthCheck;
 import org.stevenlooman.sw.magik.checks.MethodComplexityCheck;
+import org.stevenlooman.sw.magik.checks.SimplifyIfCheck;
 import org.stevenlooman.sw.magik.checks.SizeZeroEmptyCheck;
 import org.stevenlooman.sw.magik.checks.SyntaxErrorCheck;
 import org.stevenlooman.sw.magik.checks.TrailingWhitespaceCheck;
@@ -41,6 +42,7 @@ public final class CheckList {
       LineLengthCheck.class,
       MethodComplexityCheck.class,
       SyntaxErrorCheck.class,
+      SimplifyIfCheck.class,
       SizeZeroEmptyCheck.class,
       TrailingWhitespaceCheck.class,
       UnusedVariableCheck.class,