+++ CHANGES.md
@@ -3,6 +3,8 @@ Changes
 
 0.1.4 (unreleased)
 
+- Fixes after reducing size of magik-lint jar
+
 
 0.1.3 (2019-07-07)