+++ CHANGES.md
@@ -7,6 +7,7 @@ Changes
 - Rename ParserErrorCheck to SyntaxErrorCheck
 - Make LineLengthCheck a minor infraction and signal from nth column
 - Better handle multiple disabled checks
+- Add FormattingCheck
 
 
 0.1.3 (2019-07-07)