+++ CHANGES.md
@@ -10,6 +10,7 @@ Changes
 - Add FormattingCheck
 - Add ForbiddenCallCheck
 - Make magik-lint output 0-base column numbers
+- Add SimplifyIfCheck
 
 
 0.1.3 (2019-07-07)